requirements.txt
flask>=2.3.0
requests>=2.31.0
flask-cors>=4.0.0
rdflib>=6.3.2
pyshacl>=0.25.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
SPARQLWrapper>=2.0.0
